Mainline merge as of 2003-05-11
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/tree-ssa-20020619-branch@66748 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/c-common.h
@@ -1125,11 +1125,11 @@ enum c_tree_code {
    langhook should take care of initialization of this array.  */
 extern bool statement_code_p[MAX_TREE_CODES];
 
-#define STATEMENT_CODE_P(CODE) statement_code_p[(int) (CODE)]
+#define STATEMENT_CODE_P(CODE) statement_code_p[(size_t) (CODE)]
 
 #define INIT_STATEMENT_CODES(STMT_CODES)			\
   do {								\
-    unsigned int i;						\
+    size_t i;						\
     memset (&statement_code_p, 0, sizeof (statement_code_p));	\
     for (i = 0; i < ARRAY_SIZE (STMT_CODES); i++)		\
       statement_code_p[STMT_CODES[i]] = true;			\
